pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "markdown_sanitize"
version = "0.1.0"
description = "Sanitize Markdown statements for safe storage and later rendering."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Daniel Correa lobato", email = "daniel@lobato.org" }
]
requires-python = ">=3.10"
dependencies = [
  "markdown-it-py>=3.0,<4.0",
]
keywords = ["markdown", "sanitize", "sanitizer", "statements"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Text Processing :: Markup",
]

[project.urls]
Homepage = "https://github.com/dclobato/markdown_sanitize"
Repository = "https://github.com/dclobato/markdown_sanitize"
Issues = "https://github.com/dclobato/markdown_sanitize/issues"

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.11",
  "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/markdown_sanitize"]

[tool.hatch.envs.default.scripts]
test = "pytest"
lint = "ruff check ."
typecheck = "mypy src"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
packages = ["markdown_sanitize"]
